Deleting and updating of image from storage

+++ app/Http/Controllers/PostController.php
@@ -110,13 +110,28 @@ class PostController extends Controller implements HasMiddleware
         Gate::authorize('modify', $post);
 
         //Validate
-        $fields = $request->validate([
+        $request->validate([
             'title' => ['required', 'max:255'],
-            'body' => ['required']
+            'body' => ['required'],
+            'image' => ['nullable', 'file', 'max:10000' ,'mimes:png,jpg,webp']
         ]);
-        
+
+        //Store image if exists
+        $path = $post->image ?? null;
+        if($request->hasFile('image')) {
+            if($post->image) {
+                Storage::disk('public')->delete($post->image);
+            }
+
+            $path = Storage::disk('public')->put('/posts_images', $request->image);
+        } 
+
         //Update a post
-        $post->update($fields);
+        $post->update([
+            'title' => $request->title,
+            'body' => $request->body,
+            'image' => $path
+        ]);
 
         //redirect to dashboard
         return redirect('dashboard')->with('success', 'Your post was updated');
@@ -127,6 +142,13 @@ class PostController extends Controller implements HasMiddleware
      */
     public function destroy(Post $post)
     {
+
+        Gate::authorize('modify', $post);
+
+        if($post->image) {
+            Storage::disk('public')->delete($post->image);
+        }
+
         $post->delete();
 
         //Redirect
